# backend/openapi_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  description: |
    api spec for StoryNet, covering the backend API and the internal apis of the pdf_service and ner_service. all public endpoints require JWT authentication unless explicitly marked otherwise.
  title: StoryNet API
  version: 2.0.0
servers:
#- description: backend API
#  url: https://angles-server.ddns.net:5001/api
- description: Internal pdf_service (OCR) — not publicly reachable
  url: http://pdf-service.app.svc.cluster.local:5001
- description: Internal ner_service (NER/LLM) — not publicly reachable
  url: http://ner-service.app.svc.cluster.local:5000
security:
- bearerAuth: []
...
paths:
...
  /entities/search:
    get:
      description: |
        Searches the Neo4j database for FtM entities matching the given query. Results are filtered by visibility: only entities that are public OR owned by the authenticated user are returned. The `schema` parameter narrows results to a specific FtM schema type. Any combination of parameters may be used; at least one of `q` or `schema` is recommended.
      operationId: entities_search_get
      parameters:
      - description: |
          Free-text query matched against entity properties (name, aliases, identifiers, etc.). Fuzzy matching is applied when `fuzzy=true`.
        example: John Doe
        explode: true
        in: query
        name: q
        required: false
        schema:
          type: string
        style: form
      - description: "Filter by FtM schema type (e.g. Person, Organization, Ownership)."
        example: Person
        explode: true
        in: query
        name: schema
        required: false
        schema:
          type: string
        style: form
      - description: Enable fuzzy/approximate string matching (default true).
        explode: true
        in: query
        name: fuzzy
        required: false
        schema:
          default: true
          type: boolean
        style: form
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: page_size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntitySearchResponse"
          description: Paginated list of matching entities.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
      summary: Search for entities with optional fuzzy matching
      tags:
      - Entities
      x-openapi-router-controller: openapi_server.controllers.entities_controller
  /entities/{id}:
    get:
      description: |
        Returns the current live state of an entity from Neo4j. Use this endpoint when reopening a project to check whether saved entities have changed since the project snapshot was taken. Returns 404 if the entity does not exist or is not visible to the user.
      operationId: entities_id_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FtMEntity"
          description: Entity found.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found or not visible to this user.
      summary: Get a specific entity by ID
      tags:
      - Entities
      x-openapi-router-controller: openapi_server.controllers.entities_controller
  /entities/{id}/documents:
    get:
      description: |
        Returns all documents in which this entity appears via its provenance entries. Only documents that are public or owned by the authenticated user are returned.
      operationId: entities_id_documents_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: page_size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
          description: Paginated list of source documents.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found or not visible to this user.
      summary: Get documents that are the source of an entity
      tags:
      - Entities
      x-openapi-router-controller: openapi_server.controllers.entities_controller
  /entities/{id}/relationships:
    get:
      description: |
        Returns FtM interstitial entities (Ownership, Membership, Directorship, etc.) that reference this entity in any of their entity-typed properties. This is the primary graph traversal endpoint; call it recursively from the frontend to build the relationship graph. Results respect visibility rules. The `depth` parameter performs server-side recursive expansion up to the requested number of hops.
      operationId: entities_id_relationships_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: |
          Number of relationship hops to expand recursively (1 = immediate neighbours only). Values above 3 may be slow on large graphs.
        explode: true
        in: query
        name: depth
        required: false
        schema:
          default: 1
          maximum: 5
          minimum: 1
          type: integer
        style: form
      - description: Filter returned relationships by FtM schema type.
        explode: true
        in: query
        name: schema
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: page_size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntitySearchResponse"
          description: |
            Paginated list of interstitial FtM entities representing relationships. Each entry's properties contain entity-type values referencing connected node entities.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found or not visible to this user.
      summary: Get all relationships connected to an entity
      tags:
      - Entities
      x-openapi-router-controller: openapi_server.controllers.entities_controller
...
  /relationships/{id}:
    get:
      description: |
        Convenience endpoint equivalent to GET /entities/{id}, but semantically signals that the caller expects an Interval-subtype entity. Returns the same FtMEntity schema. Visibility rules apply.
      operationId: relationships_id_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FtMEntity"
          description: Relationship entity found.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found or not visible to this user.
      summary: Get a specific interstitial (relationship) entity by ID
      tags:
      - Entities
      x-openapi-router-controller: openapi_server.controllers.entities_controller
  /relationships/{id}/documents:
    get:
      description: |
        Returns all documents in which this relationship entity appears via its provenance. Visibility rules apply.
      operationId: relationships_id_documents_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: page_size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
          description: Paginated list of source documents.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found or not visible to this user.
      summary: Get documents that are the source of a relationship
      tags:
      - Entities
      x-openapi-router-controller: openapi_server.controllers.entities_controller
components:
...
  schemas:
    FtMEntity:
      example:
        schema: Person
        provenance:
        - page_number: 1
          document_id: document_id
        - page_number: 1
          document_id: document_id
        first_seen: 2000-01-23T04:56:07.000+00:00
        public: false
        confidence: 0.87
        caption: caption
        id: id
        properties:
          name:
          - John Doe
          nationality:
          - us
          - au
          birthDate:
          - "1982"
        owner_ids:
        - owner_ids
        - owner_ids
        last_changed: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Stable unique identifier (hash of canonical values).
          title: id
          type: string
        schema:
          description: |
            FtM core schema type. Node-like examples: Person, Organization, Company, LegalEntity, Asset, Vehicle, Vessel, Aircraft, Address, Passport, BankAccount. Edge-like examples: Ownership, Membership, Directorship, Family, Associate, UnknownLink, Sanction, Payment. This is not an exhaustive list; any valid FtM core schema is accepted.
          example: Person
          title: schema
          type: string
        caption:
          description: |
            Human-readable label computed from featured properties (e.g. name). Read-only; computed by the backend.
          title: caption
          type: string
        properties:
          additionalProperties:
            items:
              type: string
            type: array
          description: |
            Multi-valued property bag. ALL values are arrays of strings, even scalar-feeling fields like name or birthDate. Property names are defined by the FtM schema for this entity type. Entity-typed properties (e.g. Ownership.owner) hold the IDs of other entities.
          example:
            name:
            - John Doe
            nationality:
            - us
            - au
            birthDate:
            - "1982"
          title: properties
          type: object
        confidence:
          description: |
            NER confidence score assigned by the ner_service. Stored for all entities regardless of value. Null only for manually created entities.
          example: 0.87
          format: float
          maximum: 1.0
          minimum: 0.0
          title: confidence
          type: number
        public:
          default: false
          description: |
            True if any owning user has released this entity publicly. Derived from the public flags of all documents it was sourced from.
          title: public
          type: boolean
        owner_ids:
          description: User IDs who own at least one source document for this entity.
          items:
            type: string
          title: owner_ids
          type: array
        provenance:
          description: |
            All document+page references that contributed to this entity. An entity sourced from multiple documents will have multiple provenance entries.
          items:
            $ref: "#/components/schemas/ProvenanceEntry"
          title: provenance
          type: array
        first_seen:
          format: date-time
          readOnly: true
          title: first_seen
          type: string
        last_changed:
          format: date-time
          readOnly: true
          title: last_changed
          type: string
      required:
      - id
      - properties
      - schema
      title: FtMEntity
      type: object
    ProvenanceEntry:
      example:
        page_number: 1
        document_id: document_id
      properties:
        document_id:
          title: document_id
          type: string
        page_number:
          description: 1-indexed page number. Null for plain-text documents.
          minimum: 1
          nullable: true
          title: page_number
          type: integer
      required:
      - document_id
      title: ProvenanceEntry
      type: object
    Document:
      example:
        error_message: error_message
        entity_count: 5
        filename: filename
        public: true
        created_at: 2000-01-23T04:56:07.000+00:00
        id: id
        type: text
        owner_ids:
        - owner_ids
        - owner_ids
        status: queued
      properties:
        id:
          title: id
          type: string
        filename:
          title: filename
          type: string
        type:
          enum:
          - text
          - pdf
          title: type
          type: string
        status:
          enum:
          - queued
          - ocr_processing
          - ner_processing
          - completed
          - failed
          title: status
          type: string
        public:
          description: |
            True if one or more owners have released this document publicly. Any owner may set this to true. Once public it can only be reversed by all owners agreeing (business logic, not enforced here).
          title: public
          type: boolean
        owner_ids:
          description: IDs of all users who have uploaded this document.
          items:
            type: string
          title: owner_ids
          type: array
        entity_count:
          description: Number of FtM entities extracted from this document.
          readOnly: true
          title: entity_count
          type: integer
        created_at:
          format: date-time
          readOnly: true
          title: created_at
          type: string
        error_message:
          description: Present when status=failed.
          nullable: true
          title: error_message
          type: string
      required:
      - filename
      - id
      - owner_ids
      - public
      - status
      - type
      title: Document
      type: object
...
    EntitySearchResponse:
      example:
        total: 0
        page: 6
        results:
        - schema: Person
          provenance:
          - page_number: 1
            document_id: document_id
          - page_number: 1
            document_id: document_id
          first_seen: 2000-01-23T04:56:07.000+00:00
          public: false
          confidence: 0.87
          caption: caption
          id: id
          properties:
            name:
            - John Doe
            nationality:
            - us
            - au
            birthDate:
            - "1982"
          owner_ids:
          - owner_ids
          - owner_ids
          last_changed: 2000-01-23T04:56:07.000+00:00
        - schema: Person
          provenance:
          - page_number: 1
            document_id: document_id
          - page_number: 1
            document_id: document_id
          first_seen: 2000-01-23T04:56:07.000+00:00
          public: false
          confidence: 0.87
          caption: caption
          id: id
          properties:
            name:
            - John Doe
            nationality:
            - us
            - au
            birthDate:
            - "1982"
          owner_ids:
          - owner_ids
          - owner_ids
          last_changed: 2000-01-23T04:56:07.000+00:00
        page_size: 1
      properties:
        total:
          title: total
          type: integer
        page:
          title: page
          type: integer
        page_size:
          title: page_size
          type: integer
        results:
          items:
            $ref: "#/components/schemas/FtMEntity"
          title: results
          type: array
      title: EntitySearchResponse
      type: object
    DocumentListResponse:
      example:
        total: 0
        page: 6
        results:
        - error_message: error_message
          entity_count: 5
          filename: filename
          public: true
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
          type: text
          owner_ids:
          - owner_ids
          - owner_ids
          status: queued
        - error_message: error_message
          entity_count: 5
          filename: filename
          public: true
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
          type: text
          owner_ids:
          - owner_ids
          - owner_ids
          status: queued
        page_size: 1
      properties:
        total:
          title: total
          type: integer
        page:
          title: page
          type: integer
        page_size:
          title: page_size
          type: integer
        results:
          items:
            $ref: "#/components/schemas/Document"
          title: results
          type: array
      title: DocumentListResponse
      type: object
...
    ErrorResponse:
      example:
        details:
          key: ""
        error: ENTITY_NOT_FOUND
        message: message
      properties:
        error:
          description: Machine-readable error code.
          example: ENTITY_NOT_FOUND
          title: error
          type: string
        message:
          description: Human-readable description.
          title: message
          type: string
        details:
          additionalProperties: true
          nullable: true
          title: details
          type: object
      required:
      - error
      - message
      title: ErrorResponse
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: openapi_server.controllers.security_controller.info_from_bearerAuth